requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.0
matplotlib==3.6.2
seaborn==0.12.1
snowflake-connector-python==2.7.4
snowflake-sqlalchemy==1.3.0
jupyter==1.0.0
boto3==1.26.0
ipywidgets==7.6.5